refactor: Extract alpharat-eval-core crate with shared Backend interface
Backend trait, EvalResult, BackendError, SmartUniformBackend, and
smart_uniform_prior now live in alpharat-eval-core. Both alpharat-mcts
and alpharat-mcgs depend on it. Downstream crates (sampling, bot,
python bindings) unchanged via re-exports from alpharat-mcts.
