+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["crates/alpharat-bot", "crates/alpharat-mcgs", "crates/alpharat-mcts", "crates/alpharat-mcts-python", "crates/alpharat-sampling"]
+members = ["crates/alpharat-bot", "crates/alpharat-eval-core", "crates/alpharat-mcgs", "crates/alpharat-mcts", "crates/alpharat-mcts-python", "crates/alpharat-sampling"]
 resolver = "2"
