+++ crates/alpharat-eval-core/Cargo.toml
@@ -1,13 +1,10 @@
 [package]
-name = "alpharat-mcgs"
+name = "alpharat-eval-core"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
-name = "alpharat_mcgs"
+name = "alpharat_eval_core"
 
 [dependencies]
-alpharat-eval-core = { path = "../alpharat-eval-core" }
 pyrat = { package = "pyrat-rust", git = "https://github.com/mintiti/pyrat-rust", branch = "main", default-features = false }
-rand = { version = "0.8", features = ["small_rng"] }
-rand_distr = "0.4"
