+++ crates/alpharat-eval-core/src/test_util.rs
